Further modularize code
