+++ canvas.py
@@ -1,0 +1,19 @@
+from PyQt6.QtWidgets import QWidget, QVBoxLayout
+from matplotlib.figure import Figure
+from matplotlib.backends.backend_qt6agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.backends.backend_qt import NavigationToolbar2QT as NavigationToolbar
+
+
+class MatplotlibCanvas(QWidget):
+    """Wrapper around a Matplotlib figure for use in Qt widgets."""
+
+    def __init__(self, parent=None, width=8, height=7, dpi=100):
+        super().__init__(parent)
+        self.figure = Figure(figsize=(width, height), dpi=dpi)
+        self.axes = self.figure.add_subplot(111)
+        self.canvas = FigureCanvas(self.figure)
+        self.toolbar = NavigationToolbar(self.canvas, self)
+        layout = QVBoxLayout()
+        layout.addWidget(self.toolbar)
+        layout.addWidget(self.canvas)
+        self.setLayout(layout)
