+++ subwoofer_simulation.py
@@ -18,10 +18,6 @@ import sys
 import numpy as np
 import math
 import pandas as pd
-import matplotlib.pyplot as plt
-from matplotlib.figure import Figure
-from matplotlib.backends.backend_qt6agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.backends.backend_qt import NavigationToolbar2QT as NavigationToolbar
 import matplotlib.patches as patches
 from matplotlib.path import Path
 import matplotlib.transforms as mtransforms
@@ -34,398 +30,14 @@ from PyQt6.QtWidgets import (QApplication, QMainWindow, QWidget, QVBoxLayout, QH
 from PyQt6.QtCore import Qt, QThread, pyqtSignal, QObject
 
 import warnings
-
-# --- Parametri Globali Iniziali ---
-PARAM_RANGES = {
-    'delay_ms': (0.0, 500.0),
-    'gain_db': (-20.0, 6.0),
-    'polarity': [-1, 1], # Note: polarity is a list, not a tuple, for np.random.choice
-    'spl_rms': (70.0, 140.0),
-    'sub_width_depth': (0.1, 2.0),
-    'angle': (0.0, 2 * np.pi) # Angle should be a tuple for uniform
-}
-DEFAULT_SUB_SPL_RMS = 85.0
-DEFAULT_SUB_WIDTH = 0.4
-DEFAULT_SUB_DEPTH = 0.5
-DEFAULT_ARRAY_FREQ = 60.0
-DEFAULT_ARRAY_RADIUS = 1.0
-DEFAULT_VORTEX_MODE = 1
-DEFAULT_LINE_ARRAY_STEERING_DEG = 0.0
-DEFAULT_ARRAY_START_ANGLE_DEG = 0.0
-DEFAULT_LINE_ARRAY_COVERAGE_DEG = 0.0
-DEFAULT_VORTEX_STEERING_DEG = 0.0
-ARROW_LENGTH = 0.4
-ARRAY_INDICATOR_CONE_WIDTH_DEG = 30.0
-ARRAY_INDICATOR_RADIUS = 2.5
-
-DEFAULT_ROOM_VERTICES = [ [-15, -15], [15, -15], [15, 15], [-15, 15] ]
-P_REF_20UPA_CALC_DEFAULT = 10**(DEFAULT_SUB_SPL_RMS/20.0)
-
-DEFAULT_TARGET_AREA_VERTICES = [ [-1, -1], [1, -1], [1, 1], [-1, 1] ]
-DEFAULT_AVOIDANCE_AREA_VERTICES = [ [-4, 0], [-3, 0], [-3, 1], [-4, 1] ]
-
-DEFAULT_MAX_SPL_AVOIDANCE = 65.0
-DEFAULT_TARGET_MIN_SPL_DESIRED = 80.0
-DEFAULT_BALANCE_SLIDER_VALUE = 50 # Default for 50/50 target/avoidance balance
-
-OPTIMIZATION_MUTATION_RATE = 0.1 # Fixed mutation rate (10%)
+from constants import *
+from optimization_worker import OptimizationWorker
+from spl_calculations import calculate_spl_vectorized
+from ui_setup import UISetupMixin
 
 
-FRONT_DIRECTIVITY_BEAMWIDTH_RAD = np.pi / 6
-FRONT_DIRECTIVITY_GAIN_LIN = 10**(1.0/20.0)
-DEFAULT_SIM_SPEED_OF_SOUND = 343.0
 
-class MatplotlibCanvas(QWidget):
-    def __init__(self, parent=None, width=8, height=7, dpi=100):
-        super().__init__(parent)
-        self.figure = Figure(figsize=(width, height), dpi=dpi)
-        self.axes = self.figure.add_subplot(111)
-        self.canvas = FigureCanvas(self.figure)
-        self.toolbar = NavigationToolbar(self.canvas, self)
-        layout = QVBoxLayout(); layout.addWidget(self.toolbar); layout.addWidget(self.canvas); self.setLayout(layout)
-
-def calculate_spl_vectorized(px, py, freq, c_val, current_sorgenti_list):
-    if freq <= 0 or c_val <= 0:
-        return np.full(px.shape, -np.inf)
-
-    total_amplitude = np.zeros_like(px, dtype=np.complex128)
-    wavelength = c_val / freq
-    if wavelength == 0:
-        return np.full(px.shape, -np.inf)
-    k = 2 * np.pi / wavelength
-
-    for sub_data in current_sorgenti_list:
-        sub_x, sub_y = sub_data['x'], sub_data['y']
-        
-        distance = np.sqrt((px - sub_x)**2 + (py - sub_y)**2)
-        distance[distance < 0.01] = 0.01
-        
-        base_amplitude_attenuation = (sub_data.get('pressure_val_at_1m_relative_to_pref', 1.0) * sub_data.get('gain_lin', 1.0)) / distance
-
-        sub_orientation_angle_nord = sub_data.get('angle', 0)
-        v_sub_x = np.sin(sub_orientation_angle_nord)
-        v_sub_y = np.cos(sub_orientation_angle_nord)
-        v_point_x = px - sub_x
-        v_point_y = py - sub_y
-        dot_product = v_sub_x * v_point_x + v_sub_y * v_point_y
-        mag_point = np.sqrt(v_point_x**2 + v_point_y**2)
-        mag_point[mag_point < 1e-9] = 1e-9
-        cos_delta_angle = dot_product / mag_point
-        delta_angle = np.arccos(np.clip(cos_delta_angle, -1.0, 1.0))
-
-        directive_gain_lin = np.full(px.shape, 1.0)
-        directive_gain_lin[np.abs(delta_angle) < FRONT_DIRECTIVITY_BEAMWIDTH_RAD] = FRONT_DIRECTIVITY_GAIN_LIN
-
-        final_amplitude_component = base_amplitude_attenuation * directive_gain_lin
-        
-        phase_distance = -k * distance
-        phase_delay = -2 * np.pi * freq * (sub_data.get('delay_ms', 0.0) / 1000.0)
-        phase_polarity = np.pi if sub_data.get('polarity', 1) < 0 else 0.0
-        total_phase = phase_distance + phase_delay + phase_polarity
-        
-        total_amplitude += final_amplitude_component * np.exp(1j * total_phase)
-    
-    magnitude = np.abs(total_amplitude)
-    
-    spl = np.full(magnitude.shape, -240.0)
-    non_zero_mask = magnitude > 1e-12
-    spl[non_zero_mask] = 20 * np.log10(magnitude[non_zero_mask])
-    
-    return spl
-
-
-class OptimizationWorker(QObject):
-    status_update = pyqtSignal(str)
-    finished = pyqtSignal(object)
-    def __init__(self, criterion, optim_freq_s, optim_freq_min_val, optim_freq_max_val, optim_n_freq_val,
-                 pop_size, generations, c_val, grid_res_spl, room_pts,
-                 active_target_areas_points, active_avoidance_areas_points,
-                 max_spl_avoidance, target_min_spl_desired, balance_target_avoidance,
-                 current_sorgenti_configs,
-                 sorgenti_param_locks_list):
-        super().__init__()
-        self.criterion, self.optim_freq_s = criterion, optim_freq_s
-        self.optim_freq_min_val, self.optim_freq_max_val = optim_freq_min_val, optim_freq_max_val
-        self.optim_n_freq_val = optim_n_freq_val
-        self.pop_size, self.generations = pop_size, generations
-        self.mutation_rate_val = OPTIMIZATION_MUTATION_RATE # Fixed mutation rate
-        self.c_val, self.grid_res_spl, self.room_pts = c_val, grid_res_spl, room_pts
-        self.active_target_areas_points, self.active_avoidance_areas_points = active_target_areas_points, active_avoidance_areas_points
-        
-        self.max_spl_avoidance = max_spl_avoidance
-        self.target_min_spl_desired = target_min_spl_desired
-        self.balance_target_avoidance = balance_target_avoidance # 0-100 value from slider
-
-        self.AVOIDANCE_PENALTY_FACTOR = 1.0 # How much each dB above avoidance threshold counts
-        self.TARGET_UNDERSHOOT_PENALTY_FACTOR = 2.0 # How much each dB below target min SPL counts
-
-        self.sorgenti_configs_ref = current_sorgenti_configs
-        self.sorgenti_param_locks = sorgenti_param_locks_list
-        self._stop_requested = False
-
-    def request_stop(self):
-        self._stop_requested = True
-        self.status_update.emit("Richiesta di interruzione ottimizzazione...")
-
-    def run(self):
-        self._stop_requested = False
-        if not self.sorgenti_configs_ref:
-            self.status_update.emit("Nessun subwoofer da ottimizzare."); self.finished.emit(None); return
-        self.status_update.emit(f"Inizializzazione ottimizzazione...")
-        QApplication.processEvents()
-        population = [self._create_individual_worker() for _ in range(self.pop_size)]
-        best_overall_dsp_params, best_overall_fit = None, -float('inf')
-
-        for gen in range(self.generations):
-            if self._stop_requested: self.status_update.emit("Ottimizzazione interrotta."); self.finished.emit(best_overall_dsp_params); return
-            self.status_update.emit(f"Generazione {gen+1}/{self.generations} - Calcolo Fitness...")
-            QApplication.processEvents()
-            fit_vals = [self._calculate_fitness_worker(ind) for ind in population]
-            if not fit_vals: self.status_update.emit(f"Gen {gen+1}: Fitness non calcolabile."); break
-            current_best_idx = np.argmax(fit_vals) if fit_vals else -1
-            if current_best_idx != -1 and fit_vals[current_best_idx] > best_overall_fit:
-                best_overall_fit = fit_vals[current_best_idx]; best_overall_dsp_params = [p.copy() for p in population[current_best_idx]]
-            if not fit_vals or all(f == -float('inf') for f in fit_vals): self.status_update.emit(f"Gen {gen+1}: Nessun individuo valido."); break
-            
-            valid_indices = [i for i, f in enumerate(fit_vals) if f > -float('inf')]
-            if not valid_indices: self.status_update.emit(f"Gen {gen+1}: Nessuna fitness valida."); break
-            
-            sorted_indices = sorted(valid_indices, key=lambda i: fit_vals[i], reverse=True)
-            elite_size = max(1, int(0.1 * self.pop_size))
-            new_population = [population[i] for i in sorted_indices[:elite_size]]
-            
-            tourn_s = 3
-            while len(new_population) < self.pop_size:
-                parents = []
-                for _ in range(2):
-                    competitors_indices = np.random.choice(sorted_indices, size=min(tourn_s, len(sorted_indices)), replace=False)
-                    winner_idx = competitors_indices[np.argmax([fit_vals[i] for i in competitors_indices])]
-                    parents.append(population[winner_idx])
-                
-                child1, child2 = [], []
-                # Crossover logic needs to handle each sub's DSP parameters individually
-                for p1_sub, p2_sub in zip(parents[0], parents[1]):
-                    c1_sub, c2_sub = {}, {}
-                    # Always copy the original sub_idx for tracking locks
-                    c1_sub['sub_idx_original'] = p1_sub['sub_idx_original']
-                    c2_sub['sub_idx_original'] = p2_sub['sub_idx_original']
-
-                    for key in p1_sub:
-                        if key == 'sub_idx_original': # Skip the tracking index
-                            continue
-                        
-                        # Apply crossover for actual DSP parameters
-                        if np.random.rand() < 0.5:
-                            c1_sub[key], c2_sub[key] = p1_sub[key], p2_sub[key]
-                        else:
-                            c1_sub[key], c2_sub[key] = p2_sub[key], p1_sub[key]
-                        
-                        # Recalculate gain_lin if gain_db was crossed over
-                        if key == 'gain_db':
-                            c1_sub['gain_lin'] = 10**(c1_sub['gain_db']/20.0)
-                            c2_sub['gain_lin'] = 10**(c2_sub['gain_db']/20.0)
-                    child1.append(c1_sub)
-                    child2.append(c2_sub)
-                
-                new_population.append(self._mutate_worker(child1));
-                if len(new_population) < self.pop_size: new_population.append(self._mutate_worker(child2))
-            population = new_population
-
-        final_fitness_display = f"{best_overall_fit:.3f}" if best_overall_fit > -float('inf') else "N/A"
-        final_message = f"Terminata. Miglior fitness: {final_fitness_display}."
-        self.status_update.emit(final_message); self.finished.emit(best_overall_dsp_params)
-
-    def _create_individual_worker(self):
-        individual = []
-        for i, sub_ref in enumerate(self.sorgenti_configs_ref):
-            locks = self.sorgenti_param_locks[i]
-            dsp = {'sub_idx_original': i} # Store original index for lock lookup
-            
-            # Angle
-            if not locks.get('angle', False):
-                dsp['angle'] = np.random.uniform(*PARAM_RANGES['angle'])
-            else:
-                dsp['angle'] = sub_ref['angle']
-
-            # Delay
-            if not locks.get('delay', False):
-                dsp['delay_ms'] = np.random.uniform(*PARAM_RANGES['delay_ms'])
-            else:
-                dsp['delay_ms'] = sub_ref['delay_ms']
-            
-            # Gain
-            if not locks.get('gain', False):
-                dsp['gain_db'] = np.random.uniform(*PARAM_RANGES['gain_db'])
-            else:
-                dsp['gain_db'] = sub_ref['gain_db']
-            
-            # Polarity
-            if not locks.get('polarity', False):
-                dsp['polarity'] = np.random.choice(PARAM_RANGES['polarity'])
-            else:
-                dsp['polarity'] = sub_ref['polarity']
-            
-            dsp['gain_lin'] = 10**(dsp['gain_db']/20.0)
-            individual.append(dsp)
-        return individual
-
-    def _mutate_worker(self, individual):
-        mutated = []
-        rate = self.mutation_rate_val
-        for i, sub_dsp in enumerate(individual):
-            new_dsp = sub_dsp.copy()
-            locks = self.sorgenti_param_locks[sub_dsp['sub_idx_original']] # Use original index for locks
-            
-            if not locks.get('delay', False) and np.random.rand() < rate: 
-                new_dsp['delay_ms'] = np.clip(new_dsp['delay_ms'] + np.random.normal(0, 5), *PARAM_RANGES['delay_ms'])
-            
-            if not locks.get('gain', False) and np.random.rand() < rate: 
-                new_dsp['gain_db'] = np.clip(new_dsp['gain_db'] + np.random.normal(0, 1), *PARAM_RANGES['gain_db'])
-                new_dsp['gain_lin'] = 10**(new_dsp['gain_db']/20.0)
-            
-            if not locks.get('polarity', False) and np.random.rand() < rate: 
-                new_dsp['polarity'] *= -1
-            
-            if not locks.get('angle', False) and np.random.rand() < rate: # Mutate angle if not locked
-                new_dsp['angle'] = (new_dsp['angle'] + np.random.normal(0, np.radians(5))) % (2 * np.pi) # Smaller mutation step for angle
-                
-            mutated.append(new_dsp)
-        return mutated
-
-    def _calculate_spl_map_for_fitness_evaluation(self, individual_dsp_params, freq, area_points, is_avoidance=False):
-        configs = []
-        for i, sub_ref in enumerate(self.sorgenti_configs_ref):
-            conf = sub_ref.copy()
-            # Find the DSP parameters for the current subwoofer in the individual
-            # Use original index for robust lookup
-            dsp_params_for_this_sub = next((dsp for dsp in individual_dsp_params if dsp.get('sub_idx_original') == i), None)
-            if dsp_params_for_this_sub:
-                # Apply only relevant DSP parameters (angle, delay_ms, gain_db, polarity, gain_lin)
-                conf['angle'] = dsp_params_for_this_sub.get('angle', conf['angle'])
-                conf['delay_ms'] = dsp_params_for_this_sub.get('delay_ms', conf['delay_ms'])
-                conf['gain_db'] = dsp_params_for_this_sub.get('gain_db', conf['gain_db'])
-                conf['gain_lin'] = dsp_params_for_this_sub.get('gain_lin', conf['gain_lin'])
-                conf['polarity'] = dsp_params_for_this_sub.get('polarity', conf['polarity'])
-            
-            configs.append(conf)
-        
-        if not area_points:
-            room_points_list = [p['pos'] for p in self.room_pts]
-            if not is_avoidance: area_points = [room_points_list]
-            else: return None, None
-
-        min_x=min(p[0] for points in area_points for p in points); max_x=max(p[0] for points in area_points for p in points)
-        min_y=min(p[1] for points in area_points for p in points); max_y=max(p[1] for points in area_points for p in points)
-        x_c=np.arange(min_x,max_x+self.grid_res_spl,self.grid_res_spl); y_c=np.arange(min_y,max_y+self.grid_res_spl,self.grid_res_spl)
-        if len(x_c)<2 or len(y_c)<2: return None,None
-        
-        X,Y=np.meshgrid(x_c,y_c); points_check = np.vstack((X.ravel(), Y.ravel())).T
-        mask = np.zeros_like(X, dtype=bool)
-        for points in area_points:
-            if len(points) >= 3: mask |= Path(points).contains_points(points_check).reshape(X.shape)
-        
-        if not np.any(mask): return None, None
-        
-        spl_map = np.full(X.shape, np.nan)
-        points_to_calc_x = X[mask]
-        points_to_calc_y = Y[mask]
-        
-        if freq is None or self.c_val is None:
-            return np.full(X.shape, -np.inf), mask 
-
-        spl_values = calculate_spl_vectorized(points_to_calc_x, points_to_calc_y, freq, self.c_val, configs)
-        spl_map[mask] = spl_values
-        
-        return spl_map, mask
-
-    def _calculate_fitness_worker(self, individual):
-        score_target_component = 0.0 # Base per il contributo del target
-        score_avoidance_penalty = 0.0 # Base per la penalità di evitamento
-
-        # Caso iniziale: Se non ci sono né aree target attive né aree di evitamento attive, l'ottimizzazione non ha un obiettivo.
-        if not self.active_target_areas_points and not self.active_avoidance_areas_points:
-            return -float('inf') 
-        
-        ### LOGICA FITNESS AGGIORNATA: CALCOLO AREA TARGET ###
-        if self.active_target_areas_points:
-            freq_for_level_target = self.optim_freq_s if self.criterion == 'Copertura SPL' else (self.optim_freq_min_val + self.optim_freq_max_val) / 2.0
-
-            spl_map_target, mask_target = self._calculate_spl_map_for_fitness_evaluation(
-                individual, freq_for_level_target, self.active_target_areas_points
-            )
-            
-            if spl_map_target is not None:
-                values_target = spl_map_target[mask_target & ~np.isnan(spl_map_target)]
-                
-                if values_target.size > 0:
-                    mean_target_spl = np.mean(values_target)
-                    min_target_spl = np.min(values_target)
-
-                    base_target_score = 0.0
-                    if self.criterion == 'Copertura SPL':
-                        base_target_score = mean_target_spl
-                    elif self.criterion == 'Omogeneità SPL':
-                        with warnings.catch_warnings():
-                            warnings.simplefilter("ignore", category=RuntimeWarning)
-                            freqs = np.logspace(np.log10(self.optim_freq_min_val), np.log10(self.optim_freq_max_val), self.optim_n_freq_val)
-                            maps_for_homogeneity = [self._calculate_spl_map_for_fitness_evaluation(individual, f, self.active_target_areas_points)[0] for f in freqs]
-                            
-                            valid_homogeneity_maps = [m for m in maps_for_homogeneity if m is not None and np.any(~np.isnan(m))]
-                            
-                            if valid_homogeneity_maps:
-                                stacked_homogeneity_maps = np.stack(valid_homogeneity_maps)
-                                if stacked_homogeneity_maps.shape[0] >= 2:
-                                    variations = np.nanstd(stacked_homogeneity_maps, axis=0)
-                                else:
-                                    variations = np.full(valid_homogeneity_maps[0].shape, 0.0)
-                                
-                                base_target_score = -np.nanmean(variations) if not np.all(np.isnan(variations)) else -float('inf')
-                            else:
-                                base_target_score = -float('inf')
-
-                    score_target_component = base_target_score # Fixed base weight
-
-                    if min_target_spl < self.target_min_spl_desired:
-                        undershoot_amount = (self.target_min_spl_desired - min_target_spl)
-                        score_target_component -= (undershoot_amount * self.TARGET_UNDERSHOOT_PENALTY_FACTOR)
-                else:
-                    score_target_component = -float('inf')
-            else:
-                score_target_component = -float('inf')
-
-
-        ### LOGICA FITNESS AGGIORNATA: CALCOLO AREA DI EVITAMENTO ###
-        if self.active_avoidance_areas_points:
-            freq_avoid = self.optim_freq_s if self.criterion == 'Copertura SPL' else (self.optim_freq_min_val + self.optim_freq_max_val) / 2.0
-            
-            spl_map_avoid, mask_avoid = self._calculate_spl_map_for_fitness_evaluation(
-                individual, freq_avoid, self.active_avoidance_areas_points, is_avoidance=True
-            )
-            
-            if spl_map_avoid is not None:
-                values_avoid = spl_map_avoid[mask_avoid & ~np.isnan(spl_map_avoid)]
-                
-                if values_avoid.size > 0:
-                    mean_spl_avoid = np.mean(values_avoid)
-                    if mean_spl_avoid > self.max_spl_avoidance:
-                        penalty = (mean_spl_avoid - self.max_spl_avoidance) * self.AVOIDANCE_PENALTY_FACTOR
-                        score_avoidance_penalty = penalty
-        
-        # FITNESS FINALE: Bilanciamento tra target e avoidance
-        if self.active_target_areas_points and self.active_avoidance_areas_points:
-            balance_norm = self.balance_target_avoidance / 100.0 # Normalizza da 0 a 1
-            total_fitness = (score_target_component * balance_norm) - (score_avoidance_penalty * (1 - balance_norm))
-        elif self.active_target_areas_points:
-            total_fitness = score_target_component
-        elif self.active_avoidance_areas_points:
-            total_fitness = -score_avoidance_penalty 
-        else:
-            total_fitness = -float('inf')
-            
-        return total_fitness if not np.isnan(total_fitness) else -float('inf')
-
-
-class SubwooferSimApp(QMainWindow):
+class SubwooferSimApp(QMainWindow, UISetupMixin):
     def __init__(self, parent=None):
         super().__init__(parent)
         self.setWindowTitle("Simulatore Posizionamento Subwoofer Avanzato")
@@ -495,531 +107,6 @@ class SubwooferSimApp(QMainWindow):
         self.aggiorna_ui_sub_fields()
         self.update_optim_freq_fields_visibility()
         # _update_array_ui() is now called within _setup_group_array_ui
-
-    def _setup_ui(self):
-        self.central_widget = QWidget()
-        self.setCentralWidget(self.central_widget)
-        
-        main_container_layout = QHBoxLayout(self.central_widget)
-        splitter = QSplitter(Qt.Orientation.Horizontal)
-        main_container_layout.addWidget(splitter)
-        
-        controls_scroll_area = QScrollArea()
-        controls_scroll_area.setWidgetResizable(True)
-        controls_widget = QWidget()
-        self.controls_layout = QVBoxLayout(controls_widget)
-        controls_scroll_area.setWidget(controls_widget)
-        
-        splitter.addWidget(controls_scroll_area)
-        
-        self.plot_canvas = MatplotlibCanvas(self)
-        self.ax = self.plot_canvas.axes
-        splitter.addWidget(self.plot_canvas)
-
-        splitter.setSizes([450, 1150])
-
-        group_box_style = """
-        QGroupBox {
-            border: 2px solid gray;
-            border-radius: 5px;
-            margin-top: 1ex; /* leave space at the top for the title */
-            font-weight: bold;
-        }
-        QGroupBox::title {
-            subcontrol-origin: margin;
-            subcontrol-position: top center; /* position at the top center */
-            padding: 0 3px;
-            color: lightgray;
-        }
-        """
-        self.central_widget.setStyleSheet(group_box_style)
-
-
-        self._setup_project_ui()
-        self._setup_stanza_ui()
-        self._setup_global_sub_ui()
-        self._setup_sub_config_ui()
-        self._setup_group_array_ui() # Nuova sezione unificata
-        self._setup_target_areas_ui()
-        self._setup_avoidance_areas_ui()
-        self._setup_spl_vis_ui()
-        self._setup_sim_grid_ui()
-        self._setup_optimization_ui()
-        
-        self.controls_layout.addStretch(1)
-        
-        self.status_bar = QStatusBar()
-        self.setStatusBar(self.status_bar)
-        self.status_bar.showMessage("Applicazione caricata.")
-        self.plot_canvas.canvas.mpl_connect('button_press_event', self.on_press_mpl)
-        self.plot_canvas.canvas.mpl_connect('motion_notify_event', self.on_motion_mpl)
-        self.plot_canvas.canvas.mpl_connect('button_release_event', self.on_release_mpl)
-        self.plot_canvas.canvas.mpl_connect('motion_notify_event', self.on_mouse_move_for_spl_display)
-
-
-    def _setup_project_ui(self):
-        project_group = QGroupBox("Progetto")
-        project_layout = QHBoxLayout()
-        
-        self.btn_save_project = QPushButton("Salva Progetto")
-        self.btn_save_project.clicked.connect(self.save_project_to_excel)
-        project_layout.addWidget(self.btn_save_project)
-        
-        self.btn_load_project = QPushButton("Carica Progetto")
-        self.btn_load_project.clicked.connect(self.load_project_from_excel)
-        project_layout.addWidget(self.btn_load_project)
-        
-        project_group.setLayout(project_layout)
-        self.controls_layout.addWidget(project_group)
-        
-    def _setup_stanza_ui(self):
-        stanza_group = QGroupBox("Configurazione Stanza")
-        stanza_layout = QFormLayout()
-        self.btn_add_vtx = QPushButton("Aggiungi Vertice Stanza")
-        self.btn_add_vtx.clicked.connect(self.add_stanza_vtx)
-        stanza_layout.addRow(self.btn_add_vtx)
-        self.btn_rem_vtx = QPushButton("Rimuovi Ultimo Vertice")
-        self.btn_rem_vtx.clicked.connect(self.remove_stanza_vtx)
-        stanza_layout.addRow(self.btn_rem_vtx)
-        self.selected_vtx_label = QLabel("Nessun Vertice Selezionato")
-        stanza_layout.addRow(self.selected_vtx_label)
-        self.tb_stanza_vtx_x = QLineEdit()
-        self.tb_stanza_vtx_x.setEnabled(False)
-        stanza_layout.addRow("Vertice X:", self.tb_stanza_vtx_x)
-        self.tb_stanza_vtx_y = QLineEdit()
-        self.tb_stanza_vtx_y.setEnabled(False)
-        stanza_layout.addRow("Vertice Y:", self.tb_stanza_vtx_y)
-        self.btn_update_stanza_vtx = QPushButton("Aggiorna Vertice")
-        self.btn_update_stanza_vtx.setEnabled(False)
-        self.btn_update_stanza_vtx.clicked.connect(self.on_update_selected_stanza_vertex)
-        stanza_layout.addRow(self.btn_update_stanza_vtx)
-        self.error_text_stanza_vtx_edit = QLabel("")
-        self.error_text_stanza_vtx_edit.setStyleSheet("color: red;")
-        stanza_layout.addRow(self.error_text_stanza_vtx_edit)
-        stanza_group.setLayout(stanza_layout)
-        self.controls_layout.addWidget(stanza_group)
-
-    def _setup_global_sub_ui(self):
-        global_settings_group = QGroupBox("Impostazioni Globali Subwoofer")
-        global_settings_layout = QFormLayout()
-        self.tb_global_sub_width = QLineEdit(str(self.global_sub_width))
-        global_settings_layout.addRow("Larghezza Sub (m):", self.tb_global_sub_width)
-        self.tb_global_sub_depth = QLineEdit(str(self.global_sub_depth))
-        global_settings_layout.addRow("Profondità Sub (m):", self.tb_global_sub_depth)
-        self.tb_global_sub_spl = QLineEdit(str(self.global_sub_spl_rms))
-        global_settings_layout.addRow("SPL RMS Globale (dB):", self.tb_global_sub_spl)
-        self.check_use_global_for_new = QCheckBox("Usa per Nuovi Sub Manuali")
-        self.check_use_global_for_new.toggled.connect(self.on_toggle_use_global_for_new)
-        global_settings_layout.addRow(self.check_use_global_for_new)
-        self.btn_apply_globals_to_all = QPushButton("Applica Globali a Tutti i Sub")
-        self.btn_apply_globals_to_all.clicked.connect(self.apply_global_settings_to_all_subs)
-        global_settings_layout.addRow(self.btn_apply_globals_to_all)
-        self.error_text_global_settings = QLabel("")
-        self.error_text_global_settings.setStyleSheet("color: red;")
-        global_settings_layout.addRow(self.error_text_global_settings)
-        global_settings_group.setLayout(global_settings_layout)
-        self.controls_layout.addWidget(global_settings_group)
-        
-    def _setup_sub_config_ui(self):
-        sub_group = QGroupBox("Configurazione Subwoofer")
-        sub_layout_main = QVBoxLayout()
-        sub_selector_layout = QHBoxLayout()
-        self.btn_prev_sub = QPushButton("<")
-        self.btn_prev_sub.clicked.connect(self.select_prev_sub)
-        sub_selector_layout.addWidget(self.btn_prev_sub)
-        self.sub_selector_text_widget = QLabel("Nessun Sub")
-        sub_selector_layout.addWidget(self.sub_selector_text_widget, 1, alignment=Qt.AlignmentFlag.AlignCenter)
-        self.btn_next_sub = QPushButton(">")
-        self.btn_next_sub.clicked.connect(self.select_next_sub)
-        sub_selector_layout.addWidget(self.btn_next_sub)
-        sub_layout_main.addLayout(sub_selector_layout)
-        sub_actions_layout = QHBoxLayout()
-        self.btn_add_sub = QPushButton("Aggiungi Sub")
-        self.btn_add_sub.clicked.connect(self.add_subwoofer)
-        sub_actions_layout.addWidget(self.btn_add_sub)
-        self.btn_rem_sub = QPushButton("Rimuovi Sub Corrente")
-        self.btn_rem_sub.clicked.connect(self.remove_subwoofer)
-        sub_actions_layout.addWidget(self.btn_rem_sub)
-        sub_layout_main.addLayout(sub_actions_layout)
-        sub_params_layout = QFormLayout()
-        
-        self.sub_pos_label = QLabel("X/Y Sub (m):")
-        self.tb_sub_x = QLineEdit()
-        self.tb_sub_y = QLineEdit()
-        pos_layout = QHBoxLayout()
-        pos_layout.addWidget(self.tb_sub_x)
-        pos_layout.addWidget(self.tb_sub_y)
-        sub_params_layout.addRow(self.sub_pos_label, pos_layout)
-        
-        self.tb_sub_angle = QLineEdit()
-        sub_params_layout.addRow("Angolo (°):", self.tb_sub_angle)
-        
-        self.sub_gain_label = QLabel("Trim Gain (dB):")
-        self.tb_sub_gain_db = QLineEdit()
-        sub_params_layout.addRow(self.sub_gain_label, self.tb_sub_gain_db)
-        
-        self.sub_delay_label = QLabel("Delay (ms):")
-        self.tb_sub_delay = QLineEdit()
-        sub_params_layout.addRow(self.sub_delay_label, self.tb_sub_delay)
-        
-        self.sub_polarity_label = QLabel("Polarità (+1/-1):")
-        self.tb_sub_polarity = QLineEdit()
-        sub_params_layout.addRow(self.sub_polarity_label, self.tb_sub_polarity)
-        
-        self.check_sub_angle_lock = QCheckBox("Blocca Angolo")
-        self.check_sub_angle_lock.setObjectName("angle")
-        self.check_sub_angle_lock.toggled.connect(self.on_toggle_param_lock)
-        sub_params_layout.addRow(self.check_sub_angle_lock)
-        self.check_sub_delay_lock = QCheckBox("Blocca Delay")
-        self.check_sub_delay_lock.setObjectName("delay")
-        self.check_sub_delay_lock.toggled.connect(self.on_toggle_param_lock)
-        sub_params_layout.addRow(self.check_sub_delay_lock)
-        self.check_sub_gain_lock = QCheckBox("Blocca Gain")
-        self.check_sub_gain_lock.setObjectName("gain")
-        self.check_sub_gain_lock.toggled.connect(self.on_toggle_param_lock)
-        sub_params_layout.addRow(self.check_sub_gain_lock)
-        self.check_sub_polarity_lock = QCheckBox("Blocca Polarità")
-        self.check_sub_polarity_lock.setObjectName("polarity")
-        self.check_sub_polarity_lock.toggled.connect(self.on_toggle_param_lock)
-        sub_params_layout.addRow(self.check_sub_polarity_lock)
-        self.tb_sub_width = QLineEdit()
-        sub_params_layout.addRow("Larghezza Sub (m):", self.tb_sub_width)
-        self.tb_sub_depth = QLineEdit()
-        sub_params_layout.addRow("Profondità Sub (m):", self.tb_sub_depth)
-        self.tb_sub_spl_rms = QLineEdit()
-        sub_params_layout.addRow("SPL RMS @ 1m (dB):", self.tb_sub_spl_rms)
-
-        self.btn_submit_sub_params = QPushButton("Applica Parametri")
-        self.btn_submit_sub_params.clicked.connect(self.on_submit_sub_param_qt)
-        sub_params_layout.addRow(self.btn_submit_sub_params)
-        self.error_text_sub = QLabel("")
-        self.error_text_sub.setStyleSheet("color: red;")
-        sub_params_layout.addRow(self.error_text_sub)
-        sub_layout_main.addLayout(sub_params_layout)
-        sub_group.setLayout(sub_layout_main)
-        self.controls_layout.addWidget(sub_group)
-        
-    def _setup_group_array_ui(self): # Nuova sezione unificata
-        group_array_group = QGroupBox("Gestione Gruppi e Array")
-        group_array_layout = QVBoxLayout()
-
-        # Raggruppamento Subwoofer (spostato e adattato)
-        grouping_sub_group_box = QGroupBox("Raggruppamento Manuale")
-        grouping_sub_layout = QVBoxLayout()
-        group_action_buttons_layout = QHBoxLayout()
-        self.btn_create_new_group = QPushButton("Crea Nuovo Gruppo")
-        self.btn_create_new_group.clicked.connect(self.create_new_group)
-        group_action_buttons_layout.addWidget(self.btn_create_new_group)
-        self.btn_add_to_group = QPushButton("Aggiungi a Gruppo")
-        self.btn_add_to_group.clicked.connect(self.add_sub_to_existing_group)
-        group_action_buttons_layout.addWidget(self.btn_add_to_group)
-        grouping_sub_layout.addLayout(group_action_buttons_layout)
-        group_action_buttons_layout2 = QHBoxLayout()
-        self.btn_remove_from_group = QPushButton("Rimuovi da Gruppo")
-        self.btn_remove_from_group.clicked.connect(self.remove_sub_from_group)
-        group_action_buttons_layout2.addWidget(self.btn_remove_from_group)
-        self.btn_ungroup_all = QPushButton("Sciogli Gruppo")
-        self.btn_ungroup_all.clicked.connect(self.ungroup_selected_sub_group)
-        group_action_buttons_layout2.addWidget(self.btn_ungroup_all)
-        grouping_sub_layout.addLayout(group_action_buttons_layout2)
-        self.group_status_label = QLabel("Sub selezionato: Nessun Gruppo")
-        grouping_sub_layout.addWidget(self.group_status_label)
-        self.error_text_grouping = QLabel("")
-        self.error_text_grouping.setStyleSheet("color: red;")
-        grouping_sub_layout.addWidget(self.error_text_grouping)
-        self.group_details_label = QLabel("Valori Assoluti Membri del Gruppo:")
-        self.group_details_label.setVisible(False) # Nascosto di default
-        grouping_sub_layout.addWidget(self.group_details_label)
-
-        self.group_members_list = QListWidget()
-        self.group_members_list.setMaximumHeight(120) # Limitiamo l'altezza
-        self.group_members_list.setVisible(False) # Nascosto di default
-        grouping_sub_layout.addWidget(self.group_members_list)
-        grouping_sub_group_box.setLayout(grouping_sub_layout)
-        group_array_layout.addWidget(grouping_sub_group_box)
-
-        # Creazione Array (spostato e adattato)
-        array_setup_group_box = QGroupBox("Configurazione Array")
-        self.array_setup_layout_form = QFormLayout()
-        self.array_type_combo = QComboBox()
-        self.array_type_combo.addItems(["Nessuno", "Coppia Cardioide (2 sub)", "Array End-Fire", "Array Lineare (Steering Elettrico)", "Array Vortex"])
-        self.array_type_combo.currentIndexChanged.connect(self.on_array_type_change)
-        self.array_setup_layout_form.addRow("Tipo di Array:", self.array_type_combo)
-        self.array_freq_label = QLabel("Freq. Array Design (Hz):")
-        self.array_freq_input = QLineEdit(str(DEFAULT_ARRAY_FREQ))
-        self.array_setup_layout_form.addRow(self.array_freq_label, self.array_freq_input)
-        self.array_auto_spacing_check = QCheckBox("Calcola Spaziatura/Raggio da Frequenza")
-        self.array_auto_spacing_check.toggled.connect(self.on_auto_spacing_toggle)
-        self.array_setup_layout_form.addRow(self.array_auto_spacing_check)
-        self.array_wavelength_fraction_label = QLabel("Frazione λ per Spaziatura:")
-        self.array_wavelength_fraction_combo = QComboBox()
-        self.array_wavelength_fraction_combo.addItems(["λ/4", "λ/2"])
-        self.array_setup_layout_form.addRow(self.array_wavelength_fraction_label, self.array_wavelength_fraction_combo)
-        self.array_spacing_label = QLabel("Spaziatura/Raggio (m):")
-        self.array_spacing_input = QLineEdit(str(DEFAULT_ARRAY_RADIUS))
-        self.array_setup_layout_form.addRow(self.array_spacing_label, self.array_spacing_input)
-        self.array_elements_label = QLabel("Numero Elementi:")
-        self.array_elements_input = QLineEdit("4")
-        self.array_setup_layout_form.addRow(self.array_elements_label, self.array_elements_input)
-        self.array_start_angle_label = QLabel("Orientamento/Angolo (°):")
-        self.array_start_angle_input = QLineEdit(str(DEFAULT_ARRAY_START_ANGLE_DEG))
-        self.array_setup_layout_form.addRow(self.array_start_angle_label, self.array_start_angle_input)
-        self.array_line_coverage_angle_label = QLabel("Angolo Copertura (°):")
-        self.array_line_coverage_angle_input = QLineEdit(str(DEFAULT_LINE_ARRAY_COVERAGE_DEG))
-        self.array_setup_layout_form.addRow(self.array_line_coverage_angle_label, self.array_line_coverage_angle_input)
-        self.array_line_steering_angle_label = QLabel("Angolo Steering (°):")
-        self.array_line_steering_angle_input = QLineEdit(str(DEFAULT_LINE_ARRAY_STEERING_DEG))
-        self.array_setup_layout_form.addRow(self.array_line_steering_angle_label, self.array_line_steering_angle_input)
-        self.array_vortex_mode_label = QLabel("Modalità Vortex:")
-        self.array_vortex_mode_label.setToolTip("Numero intero (es. 1, 2, -1) che definisce l'avvolgimento dell'onda sonora.\nValori più alti creano un 'nullo' di pressione più ampio al centro.")
-        self.array_vortex_mode_input = QLineEdit(str(DEFAULT_VORTEX_MODE))
-        self.array_setup_layout_form.addRow(self.array_vortex_mode_label, self.array_vortex_mode_input)
-        self.array_vortex_steering_angle_label = QLabel("Angolo Steering Vortex (°):")
-        self.array_vortex_steering_angle_input = QLineEdit(str(DEFAULT_VORTEX_STEERING_DEG))
-        self.array_setup_layout_form.addRow(self.array_vortex_steering_angle_label, self.array_vortex_steering_angle_input)
-
-        self.apply_array_config_button = QPushButton("Crea Gruppo Array")
-        self.apply_array_config_button.clicked.connect(self.apply_array_configuration)
-        self.array_setup_layout_form.addRow(self.apply_array_config_button)
-        self.array_info_label = QLabel()
-        self.array_info_label.setWordWrap(True)
-        self.array_setup_layout_form.addRow(self.array_info_label)
-        self.error_text_array_params = QLabel("")
-        self.error_text_array_params.setStyleSheet("color: red;")
-        self.array_setup_layout_form.addRow(self.error_text_array_params)
-        array_setup_group_box.setLayout(self.array_setup_layout_form)
-        group_array_layout.addWidget(array_setup_group_box)
-
-        group_array_group.setLayout(group_array_layout)
-        self.controls_layout.addWidget(group_array_group)
-        
-    def _setup_target_areas_ui(self):
-        targets_group = QGroupBox("Aree Target")
-        targets_main_layout = QVBoxLayout()
-        target_selector_layout = QHBoxLayout()
-        self.btn_prev_target_area = QPushButton("<")
-        self.btn_prev_target_area.clicked.connect(self.select_prev_target_area)
-        target_selector_layout.addWidget(self.btn_prev_target_area)
-        self.label_current_target_area = QLabel("Nessuna Area")
-        target_selector_layout.addWidget(self.label_current_target_area, 1, Qt.AlignmentFlag.AlignCenter)
-        self.btn_next_target_area = QPushButton(">")
-        self.btn_next_target_area.clicked.connect(self.select_next_target_area)
-        target_selector_layout.addWidget(self.btn_next_target_area)
-        targets_main_layout.addLayout(target_selector_layout)
-        target_actions_layout = QHBoxLayout()
-        self.btn_new_target_area = QPushButton("Nuova")
-        self.btn_new_target_area.clicked.connect(self.add_new_target_area_ui)
-        target_actions_layout.addWidget(self.btn_new_target_area)
-        self.btn_remove_selected_target_area = QPushButton("Rimuovi")
-        self.btn_remove_selected_target_area.clicked.connect(self.remove_selected_target_area_ui)
-        target_actions_layout.addWidget(self.btn_remove_selected_target_area)
-        targets_main_layout.addLayout(target_actions_layout)
-        self.check_activate_selected_target_area = QCheckBox("Attiva Area")
-        self.check_activate_selected_target_area.toggled.connect(self.toggle_selected_target_area_active)
-        targets_main_layout.addWidget(self.check_activate_selected_target_area)
-
-        self.btn_add_target_vtx = QPushButton("Aggiungi Vertice a Area Target")
-        self.btn_add_target_vtx.clicked.connect(self._add_vtx_to_current_target_area)
-        targets_main_layout.addWidget(self.btn_add_target_vtx)
-
-        self.target_vtx_list_widget = QListWidget()
-        self.target_vtx_list_widget.currentItemChanged.connect(self.on_target_vtx_selection_change)
-        self.target_vtx_list_widget.setMaximumHeight(100)
-        targets_main_layout.addWidget(self.target_vtx_list_widget)
-        
-        vtx_edit_layout = QFormLayout()
-        self.tb_target_vtx_x = QLineEdit()
-        vtx_edit_layout.addRow("Vertice X:", self.tb_target_vtx_x)
-        self.tb_target_vtx_y = QLineEdit()
-        vtx_edit_layout.addRow("Vertice Y:", self.tb_target_vtx_y)
-        self.btn_update_target_vtx = QPushButton("Aggiorna Vertice Target")
-        self.btn_update_target_vtx.clicked.connect(self.on_update_selected_target_vertex)
-        targets_main_layout.addLayout(vtx_edit_layout)
-        targets_main_layout.addWidget(self.btn_update_target_vtx)
-
-        self.error_text_target_area_mgmt = QLabel("")
-        self.error_text_target_area_mgmt.setStyleSheet("color: orange;")
-        targets_main_layout.addWidget(self.error_text_target_area_mgmt)
-        targets_group.setLayout(targets_main_layout)
-        self.controls_layout.addWidget(targets_group)
-        
-    def _setup_avoidance_areas_ui(self):
-        avoid_group = QGroupBox("Aree di Evitamento")
-        avoid_main_layout = QVBoxLayout()
-        avoid_selector_layout = QHBoxLayout()
-        self.btn_prev_avoid_area = QPushButton("<")
-        self.btn_prev_avoid_area.clicked.connect(self.select_prev_avoidance_area)
-        avoid_selector_layout.addWidget(self.btn_prev_avoid_area)
-        self.label_current_avoid_area = QLabel("Nessuna Area")
-        avoid_selector_layout.addWidget(self.label_current_avoid_area, 1, Qt.AlignmentFlag.AlignCenter)
-        self.btn_next_avoid_area = QPushButton(">")
-        self.btn_next_avoid_area.clicked.connect(self.select_next_avoidance_area)
-        avoid_selector_layout.addWidget(self.btn_next_avoid_area)
-        avoid_main_layout.addLayout(avoid_selector_layout)
-        avoid_actions_layout = QHBoxLayout()
-        self.btn_new_avoid_area = QPushButton("Nuova")
-        self.btn_new_avoid_area.clicked.connect(self.add_new_avoidance_area_ui)
-        avoid_actions_layout.addWidget(self.btn_new_avoid_area)
-        self.btn_remove_selected_avoid_area = QPushButton("Rimuovi")
-        self.btn_remove_selected_avoid_area.clicked.connect(self.remove_selected_avoidance_area_ui)
-        avoid_actions_layout.addWidget(self.btn_remove_selected_avoid_area)
-        avoid_main_layout.addLayout(avoid_actions_layout)
-        self.check_activate_selected_avoid_area = QCheckBox("Attiva Area")
-        self.check_activate_selected_avoid_area.toggled.connect(self.toggle_selected_avoidance_area_active)
-        avoid_main_layout.addWidget(self.check_activate_selected_avoid_area) # Corrected typo here
-
-        self.btn_add_avoid_vtx = QPushButton("Aggiungi Vertice a Area Evitamento")
-        self.btn_add_avoid_vtx.clicked.connect(self._add_vtx_to_current_avoidance_area)
-        avoid_main_layout.addWidget(self.btn_add_avoid_vtx)
-
-        self.avoid_vtx_list_widget = QListWidget()
-        self.avoid_vtx_list_widget.currentItemChanged.connect(self.on_avoid_vtx_selection_change)
-        self.avoid_vtx_list_widget.setMaximumHeight(100)
-        avoid_main_layout.addWidget(self.avoid_vtx_list_widget)
-
-        vtx_edit_layout = QFormLayout()
-        self.tb_avoid_vtx_x = QLineEdit()
-        vtx_edit_layout.addRow("Vertice X:", self.tb_avoid_vtx_x)
-        self.tb_avoid_vtx_y = QLineEdit()
-        vtx_edit_layout.addRow("Vertice Y:", self.tb_avoid_vtx_y)
-        self.btn_update_avoid_vtx = QPushButton("Aggiorna Vertice Avoid")
-        self.btn_update_avoid_vtx.clicked.connect(self.on_update_selected_avoid_vertex)
-        avoid_main_layout.addLayout(vtx_edit_layout)
-        avoid_main_layout.addWidget(self.btn_update_avoid_vtx)
-
-        self.error_text_avoid_area_mgmt = QLabel("")
-        self.error_text_avoid_area_mgmt.setStyleSheet("color: orange;")
-        avoid_main_layout.addWidget(self.error_text_avoid_area_mgmt)
-        avoid_group.setLayout(avoid_main_layout)
-        self.controls_layout.addWidget(avoid_group)
-
-    def _setup_spl_vis_ui(self):
-        spl_vis_group = QGroupBox("Visualizzazione SPL")
-        spl_vis_layout = QFormLayout()
-        self.slider_freq = QSlider(Qt.Orientation.Horizontal)
-        self.slider_freq.setMinimum(20)
-        self.slider_freq.setMaximum(200)
-        self.slider_freq.setValue(80)
-        self.slider_freq.valueChanged.connect(self.on_freq_change_ui_qt)
-        spl_vis_layout.addRow("Freq. Visualiz.:", self.slider_freq)
-        self.label_slider_freq_val = QLabel(f"{self.slider_freq.value()} Hz")
-        spl_vis_layout.addRow(self.label_slider_freq_val)
-        self.tb_spl_min = QLineEdit("50")
-        self.tb_spl_max = QLineEdit("100")
-        spl_vis_layout.addRow("SPL Min Display (dB):", self.tb_spl_min)
-        spl_vis_layout.addRow("SPL Max Display (dB):", self.tb_spl_max)
-        self.error_text_spl_range = QLabel("")
-        self.error_text_spl_range.setStyleSheet("color: red;")
-        spl_vis_layout.addRow(self.error_text_spl_range)
-        self.check_auto_spl_update = QCheckBox("Aggiorna SPL Automaticamente")
-        self.check_auto_spl_update.setChecked(True)
-        spl_vis_layout.addRow(self.check_auto_spl_update)
-        self.btn_update_spl = QPushButton("Aggiorna Mappa SPL")
-        self.btn_update_spl.clicked.connect(lambda: self.visualizza_mappatura_spl(self.get_slider_freq_val(), preserve_view=True))
-        spl_vis_layout.addRow(self.btn_update_spl)
-        spl_vis_group.setLayout(spl_vis_layout)
-        self.controls_layout.addWidget(spl_vis_group)
-        
-    def _setup_sim_grid_ui(self):
-        sim_grid_params_group = QGroupBox("Parametri Simulazione e Griglia")
-        sim_grid_params_layout = QFormLayout()
-        self.tb_velocita_suono = QLineEdit(str(DEFAULT_SIM_SPEED_OF_SOUND))
-        sim_grid_params_layout.addRow("Velocità Suono (m/s):", self.tb_velocita_suono)
-        self.tb_grid_res_spl = QLineEdit("0.1")
-        sim_grid_params_layout.addRow("Risoluzione Mappa SPL (m):", self.tb_grid_res_spl)
-        self.tb_grid_snap_spacing = QLineEdit(str(self.grid_snap_spacing))
-        self.tb_grid_snap_spacing.editingFinished.connect(self.update_grid_snap_params)
-        sim_grid_params_layout.addRow("Passo Griglia Snap (m):", self.tb_grid_snap_spacing)
-        self.check_grid_snap_enabled = QCheckBox("Abilita Snap Oggetti")
-        self.check_grid_snap_enabled.setChecked(self.grid_snap_enabled)
-        self.check_grid_snap_enabled.stateChanged.connect(self.update_grid_snap_params)
-        sim_grid_params_layout.addRow(self.check_grid_snap_enabled)
-        self.check_show_grid = QCheckBox("Mostra Griglia")
-        self.check_show_grid.setChecked(self.grid_show_enabled)
-        self.check_show_grid.stateChanged.connect(self.update_grid_snap_params)
-        sim_grid_params_layout.addRow(self.check_show_grid)
-        self.error_text_grid_params = QLabel("")
-        self.error_text_grid_params.setStyleSheet("color: red;")
-        sim_grid_params_layout.addRow(self.error_text_grid_params)
-        sim_grid_params_group.setLayout(sim_grid_params_layout)
-        self.controls_layout.addWidget(sim_grid_params_group)
-
-    def _setup_optimization_ui(self):
-        optim_group = QGroupBox("Ottimizzazione Automatica")
-        optim_main_layout = QVBoxLayout()
-        self.radio_btn_group_crit = QButtonGroup(self)
-        crit_layout = QHBoxLayout()
-        self.radio_copertura = QRadioButton("Copertura SPL")
-        self.radio_copertura.setChecked(True)
-        self.radio_copertura.toggled.connect(lambda: self.update_optim_freq_fields_visibility("Copertura SPL"))
-        crit_layout.addWidget(self.radio_copertura)
-        self.radio_omogeneita = QRadioButton("Omogeneità SPL")
-        self.radio_omogeneita.toggled.connect(lambda: self.update_optim_freq_fields_visibility("Omogeneità SPL"))
-        crit_layout.addWidget(self.radio_omogeneita)
-        self.radio_btn_group_crit.addButton(self.radio_copertura)
-        self.radio_btn_group_crit.addButton(self.radio_omogeneita)
-        optim_main_layout.addLayout(crit_layout)
-        optim_freq_layout = QFormLayout()
-        self.label_opt_freq_single_widget = QLabel("Freq. Ottim. (Hz):")
-        self.tb_opt_freq_single = QLineEdit("80")
-        optim_freq_layout.addRow(self.label_opt_freq_single_widget, self.tb_opt_freq_single)
-        self.label_opt_freq_min_widget = QLabel("Freq. Min (Hz):")
-        self.tb_opt_freq_min = QLineEdit("40")
-        optim_freq_layout.addRow(self.label_opt_freq_min_widget, self.tb_opt_freq_min)
-        self.label_opt_freq_max_widget = QLabel("Freq. Max (Hz):")
-        self.tb_opt_freq_max = QLineEdit("120")
-        optim_freq_layout.addRow(self.label_opt_freq_max_widget, self.tb_opt_freq_max)
-        self.label_opt_n_freq_widget = QLabel("Nr. Punti Freq.:")
-        self.tb_opt_n_freq = QLineEdit("10")
-        optim_freq_layout.addRow(self.label_opt_n_freq_widget, self.tb_opt_n_freq)
-        optim_main_layout.addLayout(optim_freq_layout)
-        self.error_text_optim_freq = QLabel("")
-        self.error_text_optim_freq.setStyleSheet("color: red;")
-        optim_main_layout.addWidget(self.error_text_optim_freq)
-        optim_algo_layout = QFormLayout()
-        self.tb_opt_pop_size = QLineEdit("50")
-        optim_algo_layout.addRow("Config. Testate:", self.tb_opt_pop_size)
-        self.tb_opt_generations = QLineEdit("30")
-        optim_algo_layout.addRow("Cicli Ottim.:", self.tb_opt_generations)
-        
-        self.tb_target_min_spl_desired = QLineEdit(str(self.target_min_spl_desired_ui_val))
-        optim_algo_layout.addRow("Min SPL Target Desiderato (dB):", self.tb_target_min_spl_desired)
-        
-        self.tb_max_spl_avoid = QLineEdit(str(self.max_spl_avoidance_ui_val))
-        optim_algo_layout.addRow("Max SPL Evit. (dB):", self.tb_max_spl_avoid)
-        
-        self.label_balance_slider = QLabel("Bilanciamento Target / Evitamento:")
-        optim_algo_layout.addRow(self.label_balance_slider)
-        self.slider_balance = QSlider(Qt.Orientation.Horizontal)
-        self.slider_balance.setMinimum(0)
-        self.slider_balance.setMaximum(100)
-        self.slider_balance.setValue(self.balance_slider_ui_val)
-        self.label_balance_value = QLabel(f"{self.slider_balance.value()}% Target")
-        self.slider_balance.valueChanged.connect(lambda val: self.label_balance_value.setText(f"{val}% Target"))
-        optim_algo_layout.addRow(self.slider_balance)
-        optim_algo_layout.addRow(self.label_balance_value)
-
-        optim_main_layout.addLayout(optim_algo_layout)
-        self.error_text_optim_params = QLabel("")
-        self.error_text_optim_params.setStyleSheet("color: red;")
-        optim_main_layout.addWidget(self.error_text_optim_params)
-        optim_buttons_layout_H1 = QHBoxLayout()
-        self.btn_optimize_widget = QPushButton("Avvia Ottimizzazione DSP")
-        self.btn_optimize_widget.clicked.connect(self.avvia_ottimizzazione_ui_qt)
-        optim_buttons_layout_H1.addWidget(self.btn_optimize_widget)
-        self.btn_stop_optimize_widget = QPushButton("Ferma")
-        self.btn_stop_optimize_widget.clicked.connect(self.stop_ottimizzazione_ui_qt)
-        self.btn_stop_optimize_widget.setEnabled(False)
-        optim_buttons_layout_H1.addWidget(self.btn_stop_optimize_widget)
-        optim_main_layout.addLayout(optim_buttons_layout_H1)
-        self.btn_unlock_dsp_params = QPushButton("Sblocca DSP per Ottim.")
-        self.btn_unlock_dsp_params.clicked.connect(self.unlock_dsp_for_optimization)
-        optim_main_layout.addWidget(self.btn_unlock_dsp_params)
-        self.status_text_optim = QLabel("Pronto.")
-        self.status_text_optim.setWordWrap(True)
-        optim_main_layout.addWidget(self.status_text_optim)
-        optim_group.setLayout(optim_main_layout)
-        self.controls_layout.addWidget(optim_group)
 
     def disegna_subwoofer_e_elementi(self):
         for i, sub_ in enumerate(self.sorgenti):
